==> et/4.results.ipynb <==
import pandas as pd
from sklearn.metrics import mean_squared_error, r2_score

def results(feature_type):
        
    pred = pd.read_pickle(f'results/{feature_type}_test_pred.pkl')
    true = pd.read_pickle(f'results/{feature_type}_test_true.pkl')
    
    print(mean_squared_error(true, pred))
    
    print(r2_score(y_true=true, y_pred=pred))

    tr = pd.read_pickle(f'cl_features/{feature_type}_train_index.pkl')
    # res.append([feature_type, mean_squared_error(true, pred)**.5, r2_score(y_true=true, y_pred=pred)])
 
    return  feature_type, mean_squared_error(true, pred)**.5, r2_score(y_true=true, y_pred=pred), len(tr)

for feature_type in ['copy_number', 'multiplier_copy_number',
                     'proteomics', 'multiplier_proteomics',
                     'transcriptomics', 'multiplier_transcriptomics'
                    ]:
    # res.append(results(feature_type))
    tr = pd.read_pickle(f'cl_features/{feature_type}_train_index.pkl')
    print(feature_type, len(tr))

res=[]
for feature_type in ['copy_number', 'multiplier_copy_number',
                     'proteomics', 'multiplier_proteomics',
                     'transcriptomics', 'multiplier_transcriptomics'
                    ]:
      
    res.append(results(feature_type))
    

res = pd.DataFrame(res, columns=['feature_type', 'rmse', 'r2', 'n_train_data'])

#### AUC prediction accuracies

res
